pyproject.toml
[tool.poetry]
name = "my_hackathon_bot"
version = "0.1.0"
description = ""
authors = ["DevBoost1"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
telebot = "^0.0.5"
ipython = "^8.26.0"
telegram = "^0.0.1"
pymongo = "^4.8.0"
google-generativeai = "^0.7.2"
fpdf = "^1.7.2"
python-dateutil = "^2.9.0.post0"
google-search-results = "^2.4.2"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"